commented the rest of the code
Hi, this time i hashtagged (#comment) the entirety of the rest of the code i am not working on, as I am doing many examples and tries in the proposed material (book). Some of the names of variables repeat commonly, so I am test printing a lot to see what is the current output and which part of the code doesnt work as expected.