adding feature json (corrections)
trying to delete some files and edit the rest into the working condition


* updated readme to current state of affairs

* adding feature json

this is me trying to add feature "json" to an existing task. I did however leave the main file "procedure_example" to still work with the command line input.  Depending on the task it can be quickly changed into something more apropriate (input) if needed.